NCSD-2257: Upgrade to UseDotNet@2
The older version of this task was downloading the 2.2.402 dotnet core hosted image to the build server, even when it was already installed on there. Even worse, this version didn't function as expected.

This update tells the UseDotNet task to consider at the globally installed dotnet core version before downloading one to it's cache